requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu==1.5.0
keras==2.0.8
opencv-python
h5py==2.10.0
imgaug
pycocotools
IPython[all]
albumentations==0.4.5
segmentation-models-pytorch==0.1.0
catalyst==20.5.1
labelme
ipdb
yacs

# only for training
slidingwindow